pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "django-neurobank"
version = "0.10.11"
description = "A Django-based registry for the neurobank system"
readme = "README.rst"
requires-python = ">=3.10"
license = {text = "BSD 3-Clause License"}
authors = [
    {name = "C Daniel Meliza", email = "dan@meliza.org"},
]
maintainers = [
    {name = "C Daniel Meliza", email = "dan@meliza.org"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Intended Audience :: Science/Research", 
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "django>=4.2.17",
    "djangorestframework>=3.15.2",
    "djangorestframework-link-header-pagination",
    "django-filter>=22.1",
    "django-sendfile2>=0.6.0",
    "basehash>=3.0.10",
]

[project.urls]
Homepage = "https://github.com/melizalab/django-neurobank"

[dependency-groups]
dev = [
    "psycopg2-binary>=2.9.10,<3",
    "pytest-cov>=5.0.0",
    "pytest>=8.3.3,<9",
    "pytest-django>=4.9.0",
    "ruff>=0.7.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "nbank_registry.tests.settings"
django_find_project = false
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=nbank_registry --cov-report=term-missing"
testpaths = ["nbank_registry/tests"]

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py310"
extend-exclude = ["build", "attic"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]
ignore = ["E221", "E501", "E701"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.8"
ignore_missing_imports = true
strict_optional = true
check_untyped_defs = true

[tool.hatch.build]
include = ["nbank_registry/**"]
exclude = ["*test*"]
artifacts = ["README.rst"]
